refactor(Suscribe.js): Change unique suscribe form by general form
