+++ src/pages/components/Subscribe/index.js
@@ -1,21 +1,98 @@
-import React from "react"
+import React, { useState } from "react"
 import Box from '@material-ui/core/Box'
-import CircularProgress from '@material-ui/core/CircularProgress'
-import HubspotForm from 'react-hubspot-form'
+import Snackbar from '@material-ui/core/Snackbar'
+import Alert from '@material-ui/lab/Alert'
+import { useMediaQuery } from 'react-responsive'
 
-const Subscribe = () => {
-    return (
-        <HubspotForm
-            portalId='9018734'
-	        formId='df605eac-d7d1-44b5-af45-dd93d65d84ee'
-            className="hs-richtext"
-            loading={
-            <Box className="centerBox" style={{width:'100%'}}>
-                <CircularProgress style={{color:'#5484b3'}}/>
-            </Box>
+const SuscribeForm =() => {
+    const isMobile = useMediaQuery({query:'(max-width: 767px)'})
+    const [open, setOpen] = useState(false)
+    const [email, setEmail] = useState("")
+
+    const handleClose = (event, reason) => {
+        if (reason === 'clickaway') {
+        return
+        }
+
+        setOpen(false)
+    }
+
+    const handleSetEmail = (e) => {
+        setEmail(e.target.value)
+    }
+
+    const onSubmitEmailResults = async (e) => {
+        e.preventDefault()
+
+        if(email){
+        const jsonData = {
+            "fields": [
+                {
+                    "name": "email",
+                    "value": email
+                }
+            ],
+            "context": {
+                "pageUri": "https://eoscostarica.io/",
+                "pageName": "EOS Costa Rica"
+            },
+            "legalConsentOptions": {
+                "consent": {
+                    "consentToProcess": true,
+                    "text": "I agree to allow Example Company to store and process my personal data.",
+                    "communications": [
+                        {
+                            "value": true,
+                            "subscriptionTypeId": 999,
+                            "text": "I agree to receive marketing communications from Example Company."
+                        }
+                    ]
+                }
             }
-        />
-        );
-    };
+        }
+        const requestOptions = {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify(jsonData)
+        }
+        const response = await fetch('https://api.hsforms.com/submissions/v3/integration/submit/9018734/df605eac-d7d1-44b5-af45-dd93d65d84ee', requestOptions)
+        setOpen(true)
+        setEmail("")
+        }
+    }
 
-export default Subscribe;
+    return (
+        <Box>
+            <Box className="specialH3Box">
+                <h3 className="h3Footer">Subscribe to our communications </h3>
+            </Box>
+            <Box style={{paddingLeft: isMobile ? '25px' : 0, paddingRight: isMobile ? '25px' : 0}}>
+                <form onSubmit={onSubmitEmailResults}>
+                    <input
+                        className="inputFooter"
+                        type="email"
+                        id="email"
+                        name="email"
+                        placeholder="Email"
+                        value={email}
+                        onChange={handleSetEmail}
+                    />
+                    <button className="buttonPrimary" style={{padding:'11px', fontSize:'16px', float:'right'}}>Submit</button>
+                </form>
+            </Box>
+            <Box className="litleMarginTop">
+            <p style={{color:'white', fontSize:'14px'}}>
+                EOS Costa Rica respects your privacy. We need the contact information you provide to us to contact you about our services.
+                You may unsubscrie from these communications at any time.
+            </p>
+            </Box>
+            <Snackbar open={open} autoHideDuration={5000} onClose={handleClose}>
+                <Alert>
+                    Thank you for subscribing
+                </Alert>
+            </Snackbar>
+        </Box>
+    )
+}
+
+export default SuscribeForm;
